% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_cards.R
\name{deal_cards}
\alias{deal_cards}
\title{Create a new set of Texas Hold'em hands}
\usage{
deal_cards(
  n_players = 2,
  n_hands = 1,
  player_ids = NULL,
  method = c("base", "tidy", "dt")
)
}
\arguments{
\item{n_players}{numeric(). The number of players you wish to have in the game. Must be >= 2.}

\item{n_hands}{numeric() The number of hands you wish to simulate. Must be >= 1.}

\item{player_ids}{Optional. A vector of IDs for each player you wish to be in the game.}

\item{method}{character(). What R method should be used for this game: "base" (R), "tidy" (tidyverse), "dt (data.table)}
}
\value{
A completely dealt hand of Texas Hold'em in a data.frame/tibble/data.table form.
}
\description{
Simulated a hand of Texas Hold'em for a desired number of players
}
\examples{
deal_cards(n_players = 2, n_hands = 10, method = "base")
deal_cards(n_players = 10, n_hands = 2, method = "tidy")
deal_cards(n_players = 2, n_hands = 1, method = "dt")

}
