% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find Duplicates}
\usage{
find_duplicates(values, type, n_hands)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a straight in}
}
\value{
The card value (2 - 14) that defines the upper limit of a straight. If no straight exists, returns NA
}
\description{
Find Duplicates
}
