% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_flush.R
\name{find_flush}
\alias{find_flush}
\title{Find Flush Suit}
\usage{
find_flush(suits, n_hands = 0)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a flush in}
}
\value{
The suit value (1 - 4) that makes up the flush. If no flush exists, returns NA
}
\description{
Find Flush Suit
}
