% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hands.R
\name{find_flush_suit}
\alias{find_flush_suit}
\title{Find Flush Suit}
\usage{
find_flush_suit(df)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a flush in}
}
\value{
The suit value (1 - 4) that makes up the flush. If no flush exists, returns NA
}
\description{
Find Flush Suit
}
