% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hands.R
\name{find_quads}
\alias{find_quads}
\title{Find Four of a Kind Card}
\usage{
find_quads(df)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a quad in}
}
\value{
The card value (2 - 14) that makes a four of a kind. If no four of a kind exists, returns NA
}
\description{
Find Four of a Kind Card
}
