% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_straight.R
\name{find_straight}
\alias{find_straight}
\title{Find Straight High Card}
\usage{
find_straight(values, n_hands = 0)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a straight in}
}
\value{
The card value (2 - 14) that defines the upper limit of a straight. If no straight exists, returns NA
}
\description{
Find Straight High Card
}
