% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hands.R
\name{find_straight_top}
\alias{find_straight_top}
\title{Find Top of a Straight}
\usage{
find_straight_top(df)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a straight in}
}
\value{
The card value (2 - 14) that defines the upper limit of a straight. If no straight exists, returns NA
}
\description{
Find Top of a Straight
}
