% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_straightflush.R
\name{find_straightflush}
\alias{find_straightflush}
\title{Find Straight Flush High Card}
\usage{
find_straightflush(suits, values, fl_suit = NULL, st_top = NULL, n_hands = 0)
}
\arguments{
\item{suits}{card suit code}

\item{values}{card value code}

\item{fl_suit}{suit of flush}

\item{st_top}{top card of straight}

\item{n_hands}{number of hands}
}
\value{

}
\description{
Find Straight Flush High Card
}
