% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hands.R
\name{find_tripsNpairs}
\alias{find_tripsNpairs}
\title{Find Trips or Pairs}
\usage{
find_tripsNpairs(df, find = 1, top_tripple, second_dup)
}
\arguments{
\item{df}{The data frame with all hands that you want to search for a quad in}

\item{find}{Type to find. 1 looks for the highest three of a kind, 2 looks for a second three of a kind or the
highest pair, 3 looks for the second highest pair}

\item{top_tripple}{a vector of length nrow(df) that tells the function which 3 of a kind has already been found.
Required when find > 1.}

\item{second_dup}{a vector of length nrow(df) that tells the function what was the second three of a kind value or
the highest pair already found. Required when find > 2}
}
\value{
The card value (2 - 14) that makes the highest remaining trip or pair being searched for. If none exist,
         returns NA
}
\description{
If
}
