% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_make_duplicates.R
\name{help_make_duplicates}
\alias{help_make_duplicates}
\title{Help Make Duplicates}
\usage{
help_make_duplicates(cards, stage)
}
\arguments{
\item{sorted_cards}{Cards that help make a straight}
}
\value{
The index of the cards that are best given the size of the window
}
\description{
Help Make Duplicates
}
