% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_make_straight.R
\name{help_make_straight}
\alias{help_make_straight}
\title{Help Make Straight}
\usage{
help_make_straight(sorted_cards, stage, k)
}
\arguments{
\item{sorted_cards}{Cards that help make a straight}
}
\value{
The index of the cards that are best given the size of the window
}
\description{
Help Make Straight
}
