% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hand.R
\name{new_hand}
\alias{new_hand}
\title{Create a new hand of Texas Hold'em}
\usage{
new_hand(
  cards,
  chairs = NULL,
  player_ids = NULL,
  hand_id = NULL,
  auto_fill = TRUE
)
}
\arguments{
\item{cards}{A vector with all the cards available to deal in the new hand}

\item{chairs}{The number of players you wish to have in the game}

\item{player_ids}{A vector of IDs for each player you wish to be in the game}

\item{hand_id}{An ID for the hand being played}

\item{auto_fill}{If the number of chairs available is greater than length(player_ids), should we auto-generate new
players to fill all the chairs}
}
\value{
a list with two empty data.frames. players which has one row for n_players and holds each players pocket cards
         and community which has places for the 5 community cards of the game.
}
\description{
This is designed so that multiple games can be done in the same object.
}
\note{
Things we need to do/consider still:
       1) build a function that can join multiple rounds together (rbind method)
       2) Should this be turned into an explicit new class?
       3) Do we need to make a more auto-generated unique player_id - the idea being that if we want to simulate
           players in a tournament, we want to make sure they all have unique IDs. Though, we could create all the
           IDs and feed it to the function ourselves.
       4) Should we include a game_id as well? We could just do that ourselves
}
