% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TexasHoldEm.R
\name{new_TexasHoldEm_base}
\alias{new_TexasHoldEm_base}
\title{TexasHoldEm Class}
\usage{
new_TexasHoldEm_base(x)
}
\arguments{
\item{x}{data.frame, tibble, or data.table}
}
\value{
data for a round of Texas Hold'em with the appropriate type
}
\description{
TexasHoldEm Class
}
\details{
Start creating hands and games from an OOP: https://adv-r.hadley.nz/oo.html
hand_id, flop_card1, flop_card2, flop_card3, turn_card, river_card, players
   players: player_id, pocket_card1, pocket_card2, pocket_info, flop_info, turn_info, river_info
     pocket_info: num_high_pocket, suited_pocket, paired_pocket, pocket_distance, straight_pocket
     flop_info: outs, hand_rank, hand_type, hand_name
       outs: sf, fk, fh, fl, st, tk tp, pa
     turn_info: outs, hand_rank, hand_type, hand_name
       outs: sf, fk, fh, fl, st, tk tp, pa
     river_info: hand_rank, hand_type, hand_name
}
