% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outs_pa.R
\name{outs_pa}
\alias{outs_pa}
\title{Outs to make a Pair}
\usage{
outs_pa(cards, stage, output = "counts")
}
\arguments{
\item{cards}{Cards from hand, including pocket cards and all community cards}

\item{stage}{What part of deal was just performed, "flop" or "stage"?}

\item{output}{IN PROGRESS. Should function return number of outs ("counts") or the cards themselves ("cards")}
}
\value{
Either counts of cards that can make a Pair or a vector of cards that complete a Pair. If a data.frame is
input, the output is the same length as number of rows.
}
\description{
Outs to make a Pair
}
