% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outs_tk.R
\name{outs_tk}
\alias{outs_tk}
\title{Outs to make a Three of a Kind}
\usage{
outs_tk(cards, stage, output = "counts")
}
\arguments{
\item{cards}{Cards from hand, including pocket cards and all community cards}

\item{stage}{What part of deal was just performed, "flop" or "stage"?}

\item{output}{IN PROGRESS. Should function return number of outs ("counts") or the cards themselves ("cards")}
}
\value{
Either counts of cards that can make a Three of a Kind or a vector of cards that complete a Three of a Kind.
If a data.frame is input, the output is the same length as number of rows.
}
\description{
Outs to make a Three of a Kind
}
