% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_games.R
\docType{data}
\name{poker_hands}
\alias{poker_hands}
\title{Score Games of Texas Hold'em}
\format{
An object of class \code{character} of length 9.
}
\usage{
poker_hands
}
\arguments{
\item{games}{}
}
\value{

}
\description{
Score Games of Texas Hold'em
}
\keyword{datasets}
