% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_values_suits.R
\name{pull_suits}
\alias{pull_suits}
\title{Pull Suits}
\usage{
pull_suits(x)
}
\arguments{
\item{x}{A card of the form ### where the one hundreds digit represents the suit and the tens and ones digits
represent the cards value}
}
\value{
x %/% 100 which gives the 100 hundreds digit value, i.e. the suit
}
\description{
Pull Suits
}
\note{
Since it depends on how the card deck is defined, this function won't be exported
}
