% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_values_suits.R
\name{pull_value}
\alias{pull_value}
\title{Pull Values}
\usage{
pull_value(x)
}
\arguments{
\item{x}{A card of the form ### where the one hundreds digit represents the suit and the tens and ones digits
represent the cards value}
}
\value{
x %% 100 which gives the tens and ones digits, i.e. the card value
}
\description{
Pull Values
}
