% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hands.R
\name{score_all_hands}
\alias{score_all_hands}
\title{Score All Hands}
\usage{
score_all_hands(dealt_cards, stage = "river", verbose = FALSE)
}
\arguments{
\item{dealt_cards}{The cards dealt to all players in all hands}

\item{stage}{The stage where the hand is being analyzed (flop, turn, or river)}

\item{verbose}{Should R print updates as it assigns scores to individuals}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score All Hands
}
