% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hands.R
\name{score_flush}
\alias{score_flush}
\title{Score Flush}
\usage{
score_flush(suit, card_values, suit_values)
}
\arguments{
\item{suit}{The suit that makes up the flush}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Flush
}
