% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fullhouse.R
\name{score_fullhouse}
\alias{score_fullhouse}
\title{Score Full House}
\usage{
score_fullhouse(trip_card, pair_card)
}
\arguments{
\item{three}{the card value that makes up the three of a kind within a full house}

\item{two}{the card value that makes up the pair within a full house}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Full House
}
