% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pair.R
\name{score_pair}
\alias{score_pair}
\title{Score Pair}
\usage{
score_pair(pair_card, values)
}
\arguments{
\item{pair}{The value of the card that makes up the pair}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Pair
}
