% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_quad.R
\name{score_quad}
\alias{score_quad}
\title{Score Four of a Kind}
\usage{
score_quad(quad_card, values)
}
\arguments{
\item{quad_value}{The card value of the four of a kind}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Four of a Kind
}
