% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_straight.R
\name{score_straight}
\alias{score_straight}
\title{Score Straight}
\usage{
score_straight(top_card)
}
\arguments{
\item{straight_top}{The value of the high end of the straight}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Straight
}
