% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hands.R
\name{score_straight_flush}
\alias{score_straight_flush}
\title{Score Straight Flush}
\usage{
score_straight_flush(high_card, flush_suit)
}
\arguments{
\item{high_card}{The top value of the straight for a straight flush}

\item{flush_suit}{The suit of the flush for a straight flush}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Straight Flush
}
