% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_straightflush.R
\name{score_straightflush}
\alias{score_straightflush}
\title{Score Straight Flush}
\usage{
score_straightflush(top_card)
}
\arguments{
\item{high_card}{The top value of the straight for a straight flush}

\item{flush_suit}{The suit of the flush for a straight flush}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Straight Flush
}
