% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_TexasHoldEm.R
\name{score_TexasHoldEm}
\alias{score_TexasHoldEm}
\title{Score a hand or hands of Texas Hold'em}
\usage{
score_TexasHoldEm(games)
}
\arguments{
\item{games}{a dealt hand of Texas Hold'Em from deal_TexasHoldEm}
}
\value{
Each hand is given a score based on their 7 cards available
}
\description{
# Hand Rank - Description
9 - Royal Flush (RF): Suited Ace high Straight (no tie breaker)
8 - Straight Flush (SF): Suited Straight (highest card in straight breaks tie)
7 - Four of a Kind (FK): Self explanatory (highest FK value breaks tie, highest 5th card breaks subsequent tie)
6 - Full House (FH): TK and P together (highest TK value breaks tie, highest P breaks subsequent tie)
5 - Flush (F): Suited hand (highest card breaks tie, highest 2nd card breaks subsequent tie, ..., highest 5th
    card breaks subsequent tie)
4 - Straight (S): Five in a row (highest card in straight breaks tie)
3 - Three of a Kind (TK): Self explanatory (highest TK value breaks tie, highest 4th card breaks subsequent tie,
    highest 5th card breaks subsequent tie)
2 - Two Pair (TP): Self explanatory (highest pair determines win, highest 2nd pair breaks subsequent tie,
    highest 5th card breaks subsequent tie)
1 - Pair (P): Self explanatory (highest pair determines win, highest 3rd card breaks subsequent tie, highest 4th
    card breaks subsequent tie, highest 5th card breaks subsequent tie)
0 - High card (HC): Self explanatory (highest card breaks tie, highest 2nd card breaks subsequent tie, ...,
    highest 5th card breaks subsequent tie)
}
