% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hands.R
\name{score_three_of_a_kind}
\alias{score_three_of_a_kind}
\title{Score Three of a kind}
\usage{
score_three_of_a_kind(three, card_values, suit_values)
}
\arguments{
\item{three}{The value the cards that makes up the three of a kind}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Three of a kind
}
