% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hands.R
\name{score_two_pair}
\alias{score_two_pair}
\title{Score Two Pair}
\usage{
score_two_pair(twoone, twotwo, card_values, suit_values)
}
\arguments{
\item{twoone}{The value of the highest valued pair}

\item{twotwo}{The value of the lowest valued pair}

\item{card_values}{The card values of the cards in a hand}

\item{suit_values}{The suit values of the cards in a hand}
}
\value{
a data.frame with final_hand_points, final_hand, hand_type, and hand_name
}
\description{
Score Two Pair
}
