% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hand.R
\name{new_hand}
\alias{new_hand}
\title{Deal the hands for a series of Texas Hold'em games}
\usage{
new_hand(empty_table, cards, n_hands, n_players)
}
\arguments{
\item{empty_table}{An empty data.frame/tibble/data.table with a row for each hand and player. Most easily created using
the deal_cards function to ensure the format and class is correct.}

\item{cards}{a vector with every possible card code stored in card_cd. If you are playing with one deck this will have
52 rows and card_cd should uniquely identify cards. To play with multiple decks just concatenate multiple
decks.}

\item{n_hands}{The number of hands to produce}

\item{n_players}{The number of players in each hand}
}
\value{
a data frame/tibble/data.table with the original empty_table input plus unique pocket cards for each row and
          community cards that are the same for everyone.
}
\description{
A function for simulating a deal to a set number of players in a set number of rounds.
}
